pyproject.toml
[project]
name = "brh-2025-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apkit[server]>=0.3.1",
    "fastapi>=0.116.2",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.10.1",
    "pydantic-sqlite>=0.4.0",
    "pyjwt>=2.10.1",
    "tinydb>=4.8.2",
    "uvicorn>=0.36.0",
]

[dependency-groups]
dev = [
    "ruff>=0.13.1",
]

[tool.ruff]
extend-select = [
    "I",
    "FAST"
]
